Seleção de tarifa obrigatória
- Depois do usuário apertar calcular, ele verifica o valor da tarifa, podendo alterá-la
- Mudança do botão de alterar a área pro InstalacaoActivity
- Retirada dos botões de voltar de algumas telas
