+++ app/build.gradle
@@ -32,7 +32,7 @@ dependencies {
     //Add Firebase analytics
     implementation 'com.google.firebase:firebase-analytics:19.0.0'
     // Add the Firebase SDK for Crashlytics.
-    implementation 'com.google.firebase:firebase-crashlytics:18.0.1'
+    implementation 'com.google.firebase:firebase-crashlytics:18.1.0'
 
     // debugImplementation because LeakCanary should only run in debug builds. PRECISA COMENTAR O jetbrains:annotations aqui em baixo
     //debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.4'
@@ -40,11 +40,11 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'com.google.android.material:material:1.4.0'
     implementation 'com.google.firebase:firebase-core:19.0.0'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test:runner:1.3.0'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    androidTestImplementation 'androidx.test:runner:1.4.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     implementation 'org.jetbrains:annotations-java5:15.0'
 
 }
