+++ build.gradle
@@ -9,7 +9,7 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:4.1.3'
         classpath 'com.google.gms:google-services:4.3.8' //Firebase analytics
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.7.0' //Crashlytics
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.7.1' //Crashlytics
         
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
